fix warnings in the ODT plugin. can BACKPORT
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@23564 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ plugins/opendocument/imp/xp/ODi_StreamListener.cpp
@@ -386,7 +386,7 @@ void ODi_StreamListener::_handleStateAction ()
             m_currentAction = ODI_IGNORING;
             
             UT_ASSERT(m_stateAction.getElementLevel() >= -1);
-            UT_ASSERT(m_pElementStack->getStackSize() -
+            UT_ASSERT((int)m_pElementStack->getStackSize() -
                       (m_stateAction.getElementLevel()+1) >= 0);
             
             m_elemenStackSize = m_pElementStack->getStackSize() -
